src/concrete-otel/pyproject.toml
[project]
name = "concrete-otel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9, <4.0.0"
dependencies = [
    "opentelemetry-api>=1.28.2",
    "opentelemetry-exporter-otlp>=1.28.2",
    "opentelemetry-propagator-b3>=1.28.2",
    "opentelemetry-sdk>=1.28.2",
]
